rework motor valve
