+++ src/plugins/spark/features/MotorValve/MotorValveWidget.vue
@@ -1,21 +1,23 @@
 <script lang="ts">
 import { Component } from 'vue-property-decorator';
 
-import { spaceCased } from '@/helpers/functional';
 import { blockTypes } from '@/plugins/spark/block-types';
 import BlockWidgetBase from '@/plugins/spark/components/BlockWidgetBase';
 
 import { sparkStore } from '../../store';
 import { Spark3PinsBlock } from '../Spark3Pins/types';
-import { MotorValveBlock, ValveState } from './types';
+import MotorValveBasic from './MotorValveBasic.vue';
+import MotorValveFull from './MotorValveFull.vue';
+import { MotorValveBlock } from './types';
 
-@Component
+@Component({
+  components: {
+    Basic: MotorValveBasic,
+    Full: MotorValveFull,
+  },
+})
 export default class MotorValveWidget extends BlockWidgetBase {
   readonly block!: MotorValveBlock;
-
-  get valveStateName(): string {
-    return spaceCased(ValveState[this.block.data.valveState]);
-  }
 
   get pins(): Spark3PinsBlock | null {
     const block = sparkStore.blockValues(this.serviceId)
@@ -37,53 +39,31 @@ export default class MotorValveWidget extends BlockWidgetBase {
 </script>
 
 <template>
-  <q-card dark class="text-white scroll">
-    <BlockWidgetToolbar :crud="crud" />
-    <CardWarning v-if="disabled12V">
-      <template #message>
-        <span>12V is disabled.</span>
+  <GraphCardWrapper :show="inDialog">
+    <template #graph>
+      <HistoryGraph :id="widget.id" :config="graphCfg" />
+    </template>
+
+    <component :is="mode" :crud="crud" :class="cardClass">
+      <template #toolbar>
+        <component :is="toolbarComponent" :crud="crud" :mode.sync="mode" />
       </template>
-      <template #actions>
-        <q-btn text-color="white" flat label="Enable 12V" @click="enable12V" />
+
+      <template #warnings>
+        <CardWarning v-if="disabled12V">
+          <template #message>
+            <span>12V is disabled.</span>
+          </template>
+          <template #actions>
+            <q-btn text-color="white" flat label="Enable 12V" @click="enable12V" />
+          </template>
+        </CardWarning>
+        <CardWarning v-else-if="!block.data.hwDevice.id || !block.data.startChannel">
+          <template #message>
+            <span>This Motor Valve has no channel selected.</span>
+          </template>
+        </CardWarning>
       </template>
-    </CardWarning>
-    <CardWarning v-if="!block.data.hwDevice.id || !block.data.startChannel">
-      <template #message>
-        <span>This Motor Valve has no channel selected.</span>
-      </template>
-    </CardWarning>
-    <q-card-section v-else>
-      <q-item dark>
-        <q-item-section>
-          <q-item-label caption>
-            State
-          </q-item-label>
-          <DigitalStateField
-            :value="block.data.desiredState"
-            :pending="block.data.state !== block.data.desiredState"
-            :pending-reason="constrainers"
-            :disable="isDriven"
-            @input="v => { block.data.desiredState = v; saveBlock(); }"
-          />
-        </q-item-section>
-        <q-item-section>
-          <q-item-label caption>
-            Valve State
-          </q-item-label>
-          {{ valveStateName }}
-        </q-item-section>
-      </q-item>
-      <q-item dark>
-        <q-item-section>
-          <DrivenIndicator :block-id="block.id" :service-id="serviceId" />
-          <ConstraintsField
-            :value="block.data.constrainedBy"
-            :service-id="serviceId"
-            type="digital"
-            @input="v => { block.data.constrainedBy = v; saveBlock(); }"
-          />
-        </q-item-section>
-      </q-item>
-    </q-card-section>
-  </q-card>
+    </component>
+  </GraphCardWrapper>
 </template>
